BAMPFA-549: restructured objectNumber, effectiveObjectNumber to use utilities file instead
